pyproject.toml
[project]
name = "practica-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioinject>=1.10.2",
    "asyncpg>=0.31.0",
    "chromadb>=1.4.1",
    "elevenlabs>=2.35.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "langchain>=1.2.6",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-google-vertexai>=3.2.1",
    "langchain-openai>=1.1.7",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "redis[asyncio]>=7.1.0",
    "scikit-learn>=1.8.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "strawberry-graphql>=0.289.3",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]
